% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corona_outliers.R
\name{corona_outliers}
\alias{corona_outliers}
\alias{update_spec_corona_outliers}
\title{Corona outliers}
\usage{
corona_outliers(option = "ssb", freq = 12, day = "01", q_month = 1)

update_spec_corona_outliers(
  spec,
  option = "ssb",
  freq = 12,
  day = "01",
  q_month = 1,
  outlier_date_limit = "3000-01-01"
)
}
\arguments{
\item{option}{Only \code{"ssb"} implemented}

\item{freq}{frequency, \code{4} or \code{12}}

\item{day}{day of month as character}

\item{q_month}{month of quarter as \code{1}, \code{2} or \code{3}}

\item{spec}{A specification object of class "JD3_X13_SPEC" to be updated.}

\item{outlier_date_limit}{Only outliers with \code{date < outlier_date_limit} will be included in updated spec.}
}
\value{
data frame
}
\description{
Corona outliers as a type-date data frame. Also a function to update spec with these outliers.
}
\details{
Corona outliers with same date as outliers already in spec will be omitted.
}
\examples{

corona_outliers()
corona_outliers(freq = 4)

spec_a <- rjd3x13::x13_spec(name = "rsa3")
spec_a <- rjd3toolkit::set_transform(spec_a, fun = "Log")
spec_a2 <- update_spec_corona_outliers(spec_a)
as.data.frame(t(sapply(spec_a$regarima$regression$outliers,"[")))
as.data.frame(t(sapply(spec_a2$regarima$regression$outliers,"[")))

spec_b <- rjd3x13::x13_spec(name = "rsa3")
spec_b <- rjd3toolkit::set_transform(spec_b, fun = "Log")
spec_b <- rjd3toolkit::add_outlier(spec_b,type=rep("AO",3),
                                date=c("2009-01-01", "2016-01-01", "2020-05-01"))
spec_b2 <- update_spec_corona_outliers(spec_b, outlier_date_limit = "2021-11-01")
as.data.frame(t(sapply(spec_b$regarima$regression$outliers,"[")))
as.data.frame(t(sapply(spec_b2$regarima$regression$outliers,"[")))

}
