% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_ok.R
\name{crit_ok}
\alias{crit_ok}
\title{PICKMDL "first" check}
\usage{
crit_ok(sa)
}
\arguments{
\item{sa}{A \code{\link[rjd3x13]{x13}} output object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check whether \code{\link[rjd3x13]{x13}} output is ok according to
the  PICKMDL "first" method
}
\details{
Unlike \code{\link{ok}}, this function does the actual calculations.
}
\examples{

myseries <- pickmdl_data("myseries")

spec_now <- rjd3x13::x13_spec("rsa3")
spec_a <- rjd3toolkit::set_transform(spec_now,fun="Log")
a <- x13_pickmdl(myseries, spec_a)
spec_b <- rjd3toolkit::set_transform(spec_now,fun="None")
b <- x13_pickmdl(myseries, spec_b)

crit_ok(a)
crit_ok(b)

}
\seealso{
\code{\link{crit_selection}}
}
