% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_input.R
\name{filter_input}
\alias{filter_input}
\title{x13 output filters to x13 input filters}
\usage{
filter_input(sa)
}
\arguments{
\item{sa}{A \code{\link[rjd3x13]{x13}} output object}
}
\value{
list of \code{henderson.filter} (numeric) and \code{seasonal.filter} (character)
}
\description{
Elements \code{t_filter} and \code{s_filter} are transformed to input parameters \code{henderson.filter} and \code{seasonal.filter} in \code{\link[rjd3x13]{set_x11}}
}
\examples{
myseries <- pickmdl_data("myseries")

a <- rjd3x13::x13(myseries, spec = "rsa3")

a$result$decomposition$final_henderson
a$result$decomposition$final_seasonal
filter_input(a)

spec_b <- rjd3x13::x13_spec("rsa3")
spec_b <- rjd3x13::set_x11(spec_b,seasonal.filter="Stable",henderson.filter=13)
b <- rjd3x13::x13(myseries, spec = spec_b)

b$result$decomposition$final_henderson
b$result$decomposition$final_seasonal
filter_input(b)

}
