% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ok.R
\name{ok}
\alias{ok}
\title{PICKMDL information as a list}
\usage{
ok(sa)
}
\arguments{
\item{sa}{Output from \code{\link{x13_pickmdl}}}
}
\value{
List constructed from comment attribute
}
\description{
Extract \code{ok}, \code{ok_final} and \code{mdl_nr}
}
\examples{
myseries <- pickmdl_data("myseries")

spec_now <- rjd3x13::x13_spec("rsa3")
spec_a <- rjd3toolkit::set_transform(spec_now, fun = "Log")
a <- x13_pickmdl(myseries, spec_a)
spec_b <- rjd3toolkit::set_transform(spec_now, fun = "None")
b <- x13_pickmdl(myseries, spec_b)

comment(a)
comment(b)
ok(a)
ok(b)

}
\seealso{
\code{\link{crit_ok}}
}
