% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_spec_outliers.R
\name{update_spec_outliers}
\alias{update_spec_outliers}
\alias{update_outliers}
\title{Update x13 spec with outliers}
\usage{
update_spec_outliers(
  sa,
  spec = NULL,
  day = "01",
  verbose = FALSE,
  input_output = is.null(spec)
)

update_outliers(sa, spec, day = "01", null_when_no_new = TRUE, verbose = FALSE)
}
\arguments{
\item{sa}{An \code{\link[rjd3x13]{x13}} output object}

\item{spec}{An \code{\link[rjd3x13]{x13_spec}} output object}

\item{day}{Day of month as character to be used in outlier coding}

\item{verbose}{Printing information to console when \code{TRUE}.}

\item{input_output}{When \code{TRUE} output is a list of \code{x13_spec} parameters
instead of an updated spec.}

\item{null_when_no_new}{Whether to return \code{NULL} when no new outliers found.}
}
\value{
\code{update_spec_outliers} returns an updated \code{x13_spec} output object with
new outliers and updated \code{outlier.from}.
\code{update_outliers} returns a data frame with outlier variables used to update.
}
\description{
Update an \code{x13_spec} output object with outliers from an \code{x13} output object.
}
\note{
For special use, parameter \code{sa} to \code{update_outliers} can be a
data frame of outliers (as created by \code{\link{corona_outliers}}).
}
\examples{
myseries <- pickmdl_data("myseries")

spec_1 <- rjd3x13::x13_spec("rsa3")
spec_1 <- rjd3toolkit::set_transform(spec_1, fun = "Log")
spec_1 <- rjd3toolkit::set_outlier(spec_1, critical.value =3)
spec_1 <- rjd3toolkit::add_outlier(spec_1, type="AO",date="2008-09-01")


spec_2 <- rjd3toolkit::set_basic(spec_1, type="To", d1 = "2020-02-01")

a <- rjd3x13::x13(myseries, spec_2)

update_outliers(a, spec_1)

spec_3 <- update_spec_outliers(a, spec_1)

update_spec_outliers(a)

}
