% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13_both.R
\name{x13_both}
\alias{x13_both}
\title{\code{\link[rjd3x13]{x13_spec}}  and \code{\link{x13_pickmdl}} wrapped as a single function}
\usage{
x13_both(
  ts,
  spec = NULL,
  ...,
  context = NULL,
  userdefined = NULL,
  both_output = "main",
  corona = FALSE,
  pickmdl_method = "first",
  star = 1,
  when_star = warning,
  when_automdl = message,
  when_finalnotok = NULL,
  identification_end = NULL,
  identification_estimate.to = NULL,
  identify_t_filter = FALSE,
  identify_s_filter = FALSE,
  identify_outliers = TRUE,
  identify_arima_mu = TRUE,
  automdl.enabled = FALSE,
  fastfirst = TRUE,
  verbose = FALSE,
  output = "sa",
  add_comment = TRUE
)
}
\arguments{
\item{ts}{\code{x13} parameter}

\item{spec}{A "JD3_X13_SPEC" class object as generated with \code{\link[rjd3x13]{x13_spec}} or a list of several such objects as outputted from \code{\link{x13_spec_pickmdl}}.
In the case of a single object and when \code{automdl.enabled} is \code{FALSE}, \code{spec} will be converted internally
by \code{x13_spec_pickmdl} with default five arima model specifications.}

\item{...}{Specification setting on the form function__parameter. See examples.}

\item{context}{Parameter to \code{\link[rjd3x13]{x13}}. List of external regressors (calendar or other) to be used for estimation.}

\item{userdefined}{Parameter to \code{\link[rjd3x13]{x13}}  (via \code{...}  to \code{x13_pickmdl}).}

\item{both_output}{One of \code{"main"} (default, x13_pickmdl output), \code{"spec"} (spec output) or \code{"both"}.}

\item{corona}{Whether to update \code{spec} by outliers according to \code{\link{corona_outliers}}.
\code{FALSE} or \code{NULL} means no update. \code{TRUE} or \code{"ssb"} means update.}

\item{pickmdl_method}{\code{\link{crit_selection}} parameter
or one of the two extra possibilities, \code{"first_automdl"} or \code{"first_tryautomdl"}.
In both cases the \code{crit_selection} parameter is \code{"first"} and the automdl model is added as the last pickmdl model.
\itemize{
\item \strong{\code{"first_automdl"}:} The automdl model is chosen whenever no pickmdl model is ok.
In other words, the \code{star} parameter changes.
\item \strong{\code{"first_tryautomdl"}:} When no pickmdl model is ok:  The automdl model is chosen if this model is ok,
otherwise the \code{star} model is chosen.
}}

\item{star}{\code{\link{crit_selection}} parameter}

\item{when_star}{\code{\link{crit_selection}} parameter}

\item{when_automdl}{Function to be called when automdl since no pickmdl model ok. Supply NULL to do nothing.}

\item{when_finalnotok}{Function to be called, e.g. \code{\link{warning}}, when final run with final model is not ok. Supply NULL to do nothing.
See \code{\link{crit_ok}}.}

\item{identification_end}{To shorten the series before runs used to identify (arima) parameters.
That is, the series is shortened by \verb{window(ts,} \verb{end = identification_end)}.}

\item{identification_estimate.to}{To set \code{\link[rjd3toolkit]{set_estimate}} parameter \code{d1} before runs used to identify (arima) parameters.
This is an alternative to  \code{identification_end}.}

\item{identify_t_filter}{When \code{TRUE}, Henderson trend filter is identified by the shortened (see above) series.}

\item{identify_s_filter}{When \code{TRUE}, Seasonal moving average filter is identified by the shortened series.}

\item{identify_outliers}{When \code{TRUE}, Outliers are identified by the shortened series.}

\item{identify_arima_mu}{When \code{TRUE}, \code{arima.mu} is identified by the shortened series (see \code{\link{arima_mu}}).}

\item{automdl.enabled}{Logical value or any other value.
\itemize{
\item When set to \code{FALSE} (the default), the pickmdl routine is applied.
\item When set to \code{TRUE}, the automdl routine is performed.
\item For any value other than \code{TRUE} or \code{FALSE}, the ARIMA model is chosen as specified by \code{spec}.
}

Note that when \code{automdl.enabled} is not \code{FALSE}, if \code{spec} is a list containing several objects outputted from \code{x13_spec_pickmdl}, only the first object is used.}

\item{fastfirst}{When \code{TRUE} and when pickmdl with \code{crit_selection} parameter \code{"first"},
only as many models as needed are run.
This affects the output when \code{output = "all"}.}

\item{verbose}{Printing information to console when \code{TRUE}.}

\item{output}{One of \code{"sa"} (default), \code{"spec"} (final spec), \code{"sa_spec"} (both) and \code{"all"}. See examples.}

\item{add_comment}{When \code{TRUE}, a  comment attribute
(character vector with \code{ok}, \code{ok_final} and \code{mdl_nr}) will
be added to the \code{\link[rjd3x13]{x13}} output object. Use \code{\link{comment}}
to get the attribute or \code{\link{ok}} to get the attribute converted to a list.}
}
\value{
By default an \code{x13} output object, or otherwise a list as specified by parameter \code{output} and \code{both_output}.
}
\description{
Output is determined by the parameter \code{both_output}.
}
\details{
All parameters except \code{both_output} and \code{...} are parameters to \code{\link{x13_pickmdl}}.
}
\examples{

myseries <- pickmdl_data("myseries")

a <- x13_both(myseries, spec = "rsa3", set_transform__fun = "Log", verbose = TRUE)
summary(a)

# is equivalent to

spec_a <- x13_spec("rsa3")
spec_a <- rjd3toolkit::set_transform(spec_a, fun = "Log")
a2 <- x13_both(myseries,spec_a,verbose=TRUE)
a2

# several specification settings:

b <- x13_both(myseries, spec="rsa3",
  set_transform__fun = "None",
  set_easter__enable = TRUE,
  set_easter__duration = 3,
  set_easter__test = "None",
  set_outlier__outliers.type = c("LS","AO"),
  set_outlier__critical.value = 3
)

# user defined regressors and modelling context




}
