% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13_multi.R
\name{x13_multi}
\alias{x13_multi}
\title{Multiple x13 runs from multiple specifications}
\usage{
x13_multi(..., spec)
}
\arguments{
\item{...}{\code{x13} parameters}

\item{spec}{List of several \code{x13_spec} output objects. That is, \code{spec} can be output from  \code{\link{x13_spec_pickmdl}}.}
}
\value{
List of several \code{x13} output objects
}
\description{
\code{\link[rjd3x13]{x13}} is run multiple times
}
\details{
This function behaves like \code{x13} except that parameter \code{spec} is a list of multiple specifications.
}
\examples{

 myseries <- pickmdl_data("myseries")

 spec <- rjd3x13::x13_spec("rsa3")
 spec5 <- x13_spec_pickmdl(spec)

 sa5 <- x13_multi(myseries, spec = spec5)

}
