% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13_spec_pickmdl.R
\name{x13_spec_pickmdl}
\alias{x13_spec_pickmdl}
\title{Multiple X-13ARIMA model specifications}
\usage{
x13_spec_pickmdl(
  ...,
  arima.p = c(0, 0, 2, 0, 2),
  arima.d = c(1, 1, 1, 2, 1),
  arima.q = c(1, 2, 0, 2, 2),
  arima.bp = 0,
  arima.bd = 1,
  arima.bq = 1,
  automdl.enabled = FALSE
)
}
\arguments{
\item{...}{A "JD3_X13_SPEC" class object generated with \code{\link[rjd3x13]{x13_spec}}}

\item{arima.p}{'set_arima' parameters as vector.}

\item{arima.d}{'set_arima' parameters as vector.}

\item{arima.q}{'set_arima' parameters as vector.}

\item{arima.bp}{'set_arima' parameters as vector.}

\item{arima.bd}{'set_arima' parameters as vector.}

\item{arima.bq}{'set_arima' parameters as vector.}

\item{automdl.enabled}{'set_automodel' parameter}
}
\value{
List of several "JD3_X13_SPEC" class objects
}
\description{
\code{\link[rjd3x13]{x13_spec}} is run multiple times with input for multiple arima models.
}
\details{
This function behaves like \code{x13_spec} except that some of the parameters may be vectors.
These vectors must be the same length.
}
\examples{

spec <- rjd3x13::x13_spec("rsa3")
spec_list <- x13_spec_pickmdl(spec)

}
