% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13_pickmdl.R
\name{x13_pickmdl}
\alias{x13_pickmdl}
\alias{x13_automdl}
\title{x13 with PICKMDL and partial concurrent possibilities}
\usage{
x13_pickmdl(
  ts,
  spec,
  corona = FALSE,
  ...,
  pickmdl_method = "first",
  star = 1,
  when_star = warning,
  when_automdl = message,
  when_finalnotok = NULL,
  identification_end = NULL,
  identification_estimate.to = NULL,
  policy = "Outliers",
  identify_t_filter = FALSE,
  identify_s_filter = FALSE,
  identify_outliers = TRUE,
  identify_arima_mu = TRUE,
  automdl.enabled = FALSE,
  fastfirst = TRUE,
  verbose = FALSE,
  output = "sa",
  add_comment = TRUE
)

x13_automdl(..., automdl.enabled = TRUE)
}
\arguments{
\item{ts}{\code{x13} parameter}

\item{spec}{A "JD3_X13_SPEC" class object as generated with \code{\link[rjd3x13]{x13_spec}} or a list of several such objects as outputted from \code{\link{x13_spec_pickmdl}}.
In the case of a single object and when \code{automdl.enabled} is \code{FALSE}, \code{spec} will be converted internally
by \code{x13_spec_pickmdl} with default five arima model specifications.}

\item{corona}{Whether to update \code{spec} by outliers according to \code{\link{corona_outliers}}.
\code{FALSE} or \code{NULL} means no update. \code{TRUE} or \code{"ssb"} means update.}

\item{...}{Further \code{x13} parameters (currently only parameter \code{userdefined} is additional parameter to \code{x13}).}

\item{pickmdl_method}{\code{\link{crit_selection}} parameter
or one of the two extra possibilities, \code{"first_automdl"} or \code{"first_tryautomdl"}.
In both cases the \code{crit_selection} parameter is \code{"first"} and the automdl model is added as the last pickmdl model.
\itemize{
\item \strong{\code{"first_automdl"}:} The automdl model is chosen whenever no pickmdl model is ok.
In other words, the \code{star} parameter changes.
\item \strong{\code{"first_tryautomdl"}:} When no pickmdl model is ok:  The automdl model is chosen if this model is ok,
otherwise the \code{star} model is chosen.
}}

\item{star}{\code{\link{crit_selection}} parameter}

\item{when_star}{\code{\link{crit_selection}} parameter}

\item{when_automdl}{Function to be called when automdl since no pickmdl model ok. Supply NULL to do nothing.}

\item{when_finalnotok}{Function to be called, e.g. \code{\link{warning}}, when final run with final model is not ok. Supply NULL to do nothing.
See \code{\link{crit_ok}}.}

\item{identification_end}{To shorten the series before runs used to identify (arima) parameters.
That is, the series is shortened by \verb{window(ts,} \verb{end = identification_end)}.}

\item{identification_estimate.to}{To set \code{\link[rjd3toolkit]{set_estimate}} parameter \code{d1} before runs used to identify (arima) parameters.
This is an alternative to  \code{identification_end}.}

\item{policy}{Which refresh policy when model identification by shortened series. See \code{\link[rjd3x13]{x13_refresh}}.}

\item{identify_t_filter}{When \code{TRUE}, Henderson trend filter is identified by the shortened (see above) series.}

\item{identify_s_filter}{When \code{TRUE}, Seasonal moving average filter is identified by the shortened series.}

\item{identify_outliers}{When \code{TRUE}, Outliers are identified by the shortened series.}

\item{identify_arima_mu}{When \code{TRUE}, \code{arima.mu} is identified by the shortened series (see \code{\link{arima_mu}}).}

\item{automdl.enabled}{Logical value or any other value.
\itemize{
\item When set to \code{FALSE} (the default), the pickmdl routine is applied.
\item When set to \code{TRUE}, the automdl routine is performed.
\item For any value other than \code{TRUE} or \code{FALSE}, the ARIMA model is chosen as specified by \code{spec}.
}

Note that when \code{automdl.enabled} is not \code{FALSE}, if \code{spec} is a list containing several objects outputted from \code{x13_spec_pickmdl}, only the first object is used.}

\item{fastfirst}{When \code{TRUE} and when pickmdl with \code{crit_selection} parameter \code{"first"},
only as many models as needed are run.
This affects the output when \code{output = "all"}.}

\item{verbose}{Printing information to console when \code{TRUE}.}

\item{output}{One of \code{"sa"} (default), \code{"spec"} (final spec), \code{"sa_spec"} (both) and \code{"all"}. See examples.}

\item{add_comment}{When \code{TRUE}, a  comment attribute
(character vector with \code{ok}, \code{ok_final} and \code{mdl_nr}) will
be added to the \code{\link[rjd3x13]{x13}} output object. Use \code{\link{comment}}
to get the attribute or \code{\link{ok}} to get the attribute converted to a list.}
}
\value{
By default an \code{x13} output object, or otherwise a list as specified by parameter \code{output}.
}
\description{
\code{\link[rjd3x13]{x13}} can be run as usual (automdl) or with a PICKMDL specification.
The ARIMA model, outliers and filters can be identified at a certain date and then held fixed (with a new outlier-span).
}
\examples{
myseries <- pickmdl_data("myseries")

spec_a  <- rjd3x13::x13_spec(name = "rsa3")
spec_a <- rjd3toolkit::set_transform(spec_a, fun = "Log")

a <- x13_pickmdl(myseries, spec_a, verbose = TRUE)
comment(a)
ok(a)
unlist(ok(a))
summary(a$result$preprocessing)

a2 <- x13_pickmdl(myseries, spec_a, identification_end = c(2014, 2))
summary(a2$result$preprocessing)

# As above, another way
a3 <- x13_pickmdl(myseries, spec_a, identification_estimate.to = "2014-03-01")
summary(a3$result$preprocessing)

a4 <- x13_automdl(myseries, spec_a, identification_end = c(2014, 2))
summary(a4$result$preprocessing)

# As above, another way
a5 <- x13_automdl(myseries, spec_a, identification_estimate.to = "2014-03-01")
summary(a5$result$preprocessing)


allvar <- pickmdl_data("allvar")
allvar <- list(arb_dag=allvar[,1],skuddar=allvar[,2])
my_context <- modelling_context(variables=allvar)
spec_b <- rjd3x13::x13_spec(name= "rsa3")
spec_b <- rjd3toolkit::set_transform(spec_b,fun="Log")
spec_b <- rjd3toolkit::set_tradingdays(spec_b,
                       option="Userdefined",uservariable=c("r.arb_dag","r.skuddar"))
spec_b <- rjd3toolkit::set_outlier(spec_b,outliers.type=NULL)
spec_b <- rjd3toolkit::add_outlier(spec_b,type=rep("LS",20),
                           date = c("2009-01-01", "2016-01-01", "2020-03-01",
                                    "2020-04-01", "2020-05-01", "2020-06-01",
                                    "2020-07-01", "2020-08-01", "2020-09-01",
                                    "2020-10-01", "2020-11-01", "2020-12-01",
                                    "2021-01-01", "2021-02-01", "2021-03-01",
                                    "2021-04-01", "2021-05-01", "2021-06-01",
                                    "2021-07-01", "2021-08-01"))
b <- x13_pickmdl(myseries,spec_b, identification_end = c(2020, 2),context=my_context)
summary(b$result$preprocessing)

# automdl instead
b1 <- x13_automdl(myseries, spec_b, identification_end = c(2020, 2),context=my_context)
summary(b1$result$preprocessing)

# effect of identify_t_filter and identify_s_filter
set.seed(1)
rndseries <- ts(rep(1:12, 20) + (1 + (1:240)/20) * runif(240) + 0.5 * c(rep(1, 120), (1:120)^2),
                frequency = 12, start = c(2000, 1))
spec_c <- rjd3toolkit::set_outlier(rjd3x13::x13_spec("rsa3"),outliers.type=NULL)
c1 <- x13_automdl(rndseries, spec_c, identification_end = c(2009, 12))
c1
c2 <- x13_automdl(rndseries, spec_c, identification_end = c(2009, 12), identify_t_filter = TRUE)
c2
c3 <- x13_automdl(rndseries, spec_c, identification_end = c(2009, 12), identify_t_filter = TRUE,
                  identify_s_filter = TRUE)
c3


# Warning when transform.function = "None"
spec_d  <- rjd3toolkit::set_transform(rjd3x13::x13_spec("rsa3"), fun = "None")
d <- x13_pickmdl(myseries, spec_d, verbose = TRUE)

# Warning avoided (when_star) and 2nd (star) model selected
d2 <- x13_pickmdl(myseries, spec_d, star = 2, when_star = NULL, verbose = TRUE)

# automdl since no pickmdl model ok, but still not ok
d3 <- x13_pickmdl(myseries, spec_d, pickmdl_method = "first_automdl", verbose = TRUE)

# airline model (star) since automdl also not ok
d4 <- x13_pickmdl(myseries, spec_d, pickmdl_method = "first_tryautomdl", verbose = TRUE,
                  when_finalnotok = warning) # also finalnotok warning

# As b, with output = "all"
k <- x13_pickmdl(myseries, spec_b, identification_end = c(2014, 2), context = my_context,
                 output = "all", fastfirst = FALSE) # With TRUE only one model in this case
summary(k$sa$result$preprocessing)  # As summary(b$result$preprocessing)

k$mdl_nr            # index of selected model used to identify parameters
k$sa_mult[[k$mdl_nr]] # model to identify
k$crit_tab          # Table of criteria


# Effect of identify_outliers (TRUE is default)
m1 <- x13_pickmdl(myseries, rjd3toolkit::set_outlier(rjd3x13::x13_spec("rsa3"), critical.value= 3),
                  identification_end = c(2010, 2), identify_outliers = FALSE)
m2 <- x13_pickmdl(myseries, rjd3toolkit::set_outlier(rjd3x13::x13_spec("rsa3"), critical.value= 3),
                  identification_end = c(2010, 2), identify_outliers = TRUE,
                  verbose = TRUE)
m3 <- x13_pickmdl(myseries, rjd3toolkit::set_outlier(rjd3x13::x13_spec("rsa3"), critical.value= 3),
                  identification_end = c(2018, 2), identify_outliers = TRUE,
                  verbose = TRUE)



# With corona outliers (even possible when series is not long enough)
m4 <- x13_pickmdl(myseries, spec_a, verbose = TRUE, corona = TRUE)
summary(m4$result$preprocessing)
m5 <- x13_pickmdl(myseries , rjd3toolkit::set_outlier(rjd3x13::x13_spec("rsa3"), critical.value= 3),
                  identification_end = c(2010, 2), identify_outliers = TRUE,
                  verbose = TRUE, corona = TRUE)
summary(m5$result$preprocessing)


###########  quarterly series  #############

qseries <- pickmdl_data("qseries")

# Effect of identify_outliers (TRUE is default)
q1 <- x13_pickmdl(qseries, rjd3toolkit::set_outlier(rjd3x13::x13_spec("rsa3"), critical.value = 3),
                  identification_end = c(2010, 2), identify_outliers = FALSE)
q2 <- x13_pickmdl(qseries, rjd3toolkit::set_outlier(rjd3x13::x13_spec("rsa3"), critical.value = 3),
                  identification_end = c(2010, 2), identify_outliers = TRUE,
                  verbose = TRUE, output = "all")
q3 <- x13_pickmdl(qseries, q2$spec, identification_end = c(2018, 2), identify_outliers = TRUE,
                  verbose = TRUE)

# With corona outliers (even possible when series is not long enough)
q4 <- x13_pickmdl(qseries, spec_a, verbose = TRUE, corona = TRUE)
summary(q4$result$preprocessing)

q5 <- x13_pickmdl(qseries, rjd3toolkit::set_outlier(rjd3x13::x13_spec("rsa3"), critical.value = 3),
                  identification_end = c(2010, 2), identify_outliers = TRUE,
                  verbose = TRUE, corona = TRUE)
summary(q5$result$preprocessing)


}
